requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
python-multipart>=0.0.9
pillow>=10.0
numpy>=1.24
httpx>=0.27

# PaddleOCR stack (macOS uses CPU; CUDA GPU only on Linux + NVIDIA)
paddlepaddle>=2.6
paddleocr>=3.4
